Move vars from __init__ to utils/__init__
