+++ pl_bolts/datamodules/sklearn_datamodule.py
@@ -6,7 +6,7 @@ import torch
 from pytorch_lightning import LightningDataModule
 from torch.utils.data import DataLoader, Dataset
 
-from pl_bolts import _SKLEARN_AVAILABLE
+from pl_bolts.utils import _SKLEARN_AVAILABLE
 from pl_bolts.utils.warnings import warn_missing_pkg
 
 if _SKLEARN_AVAILABLE:
