+++ pl_bolts/datasets/cifar10_dataset.py
@@ -6,7 +6,7 @@ from typing import Callable, Optional, Sequence, Tuple
 import torch
 from torch import Tensor
 
-from pl_bolts import _PIL_AVAILABLE
+from pl_bolts.utils import _PIL_AVAILABLE
 from pl_bolts.datasets.base_dataset import LightDataset
 from pl_bolts.utils.warnings import warn_missing_pkg
